Intelligent Multi-Agent Email Automation System/requirements.txt
"""
Requirements file for the Intelligent Multi-Agent Email Automation System.
This file lists all the Python dependencies required for the project.
"""

# FastAPI and related packages
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7

# Email handling
python-imap>=1.0.0
email-validator>=2.0.0
aiosmtplib>=2.0.1

# NLP and ML libraries
transformers>=4.28.1
torch>=2.0.0
nltk>=3.8.1
spacy>=3.5.2

# Database and caching
motor>=3.1.2  # MongoDB async driver
redis>=4.5.4
pymongo>=4.3.3

# API integrations
google-auth>=2.17.3
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.86.0
msal>=1.22.0  # Microsoft Authentication Library

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
asyncio>=3.4.3
pyjwt>=2.6.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
